DTS-38044:Moving into release_1010 package
